Branch instructions should compare 64-bit register values

+++ utopia/src/core/mips/operator/branch.rs
@@ -1,7 +1,7 @@
 pub trait BranchOperator {
     const NAME: &'static str;
     const UNARY: bool;
-    fn apply(rs: u32, rt: u32) -> bool;
+    fn apply(rs: u64, rt: u64) -> bool;
 }
 
 pub struct Beq;
@@ -10,7 +10,7 @@ impl BranchOperator for Beq {
     const NAME: &'static str = "BEQ";
     const UNARY: bool = false;
 
-    fn apply(rs: u32, rt: u32) -> bool {
+    fn apply(rs: u64, rt: u64) -> bool {
         rs == rt
     }
 }
@@ -21,7 +21,7 @@ impl BranchOperator for Bne {
     const NAME: &'static str = "BNE";
     const UNARY: bool = false;
 
-    fn apply(rs: u32, rt: u32) -> bool {
+    fn apply(rs: u64, rt: u64) -> bool {
         rs != rt
     }
 }
@@ -32,7 +32,7 @@ impl BranchOperator for Blez {
     const NAME: &'static str = "BLEZ";
     const UNARY: bool = true;
 
-    fn apply(rs: u32, _rt: u32) -> bool {
+    fn apply(rs: u64, _rt: u64) -> bool {
         (rs as i32) <= 0
     }
 }
@@ -43,7 +43,7 @@ impl BranchOperator for Bgtz {
     const NAME: &'static str = "BGTZ";
     const UNARY: bool = true;
 
-    fn apply(rs: u32, _rt: u32) -> bool {
+    fn apply(rs: u64, _rt: u64) -> bool {
         (rs as i32) > 0
     }
 }
@@ -54,7 +54,7 @@ impl BranchOperator for Bltz {
     const NAME: &'static str = "BLTZ";
     const UNARY: bool = true;
 
-    fn apply(rs: u32, _rt: u32) -> bool {
+    fn apply(rs: u64, _rt: u64) -> bool {
         (rs as i32) < 0
     }
 }
@@ -65,7 +65,7 @@ impl BranchOperator for Bgez {
     const NAME: &'static str = "BGEZ";
     const UNARY: bool = true;
 
-    fn apply(rs: u32, _rt: u32) -> bool {
+    fn apply(rs: u64, _rt: u64) -> bool {
         (rs as i32) >= 0
     }
 }
